backend/requirements.txt
fastapi==0.115.0
uvicorn[standard]==0.31.0
pydantic==2.9.2
python-multipart==0.0.12
openai==1.54.5
python-dotenv==1.0.0
httpx==0.27.0
tenacity==9.0.0
requests==2.31.0
